PB-1140: moving constant to config file
