+++ src/api/print.api.js
@@ -20,7 +20,6 @@ import { adjustWidth } from '@/utils/styleUtils'
 
 const PRINTING_DEFAULT_POLL_INTERVAL = 2000 // interval between each polling of the printing job status (ms)
 const PRINTING_DEFAULT_POLL_TIMEOUT = 600000 // ms (10 minutes)
-export const MIN_PRINT_SCALE_SIZE = 0.0001 // when the scale is too low, the print backend can't read the exponent. So when there is a non 0 scale, we set its minimum to 0.0001
 const SERVICE_PRINT_URL = `${getViewerDedicatedServicesBaseUrl()}print3/print/mapviewer`
 const MAX_PRINT_SPEC_SIZE = 1 * 1024 * 1024 // 1MB in bytes (should be in sync with the backend)
